+++ visualization/app/codeCharta/state/metric.service.ts
@@ -1,14 +1,4 @@
-import {
-	AttributeTypes,
-	BlacklistItem,
-	BlacklistType,
-	CodeMapNode,
-	FileState,
-	MetricData,
-	AttributeType,
-	AttributeTypeValue,
-	State
-} from "../codeCharta.model"
+import { BlacklistItem, BlacklistType, CodeMapNode, FileState, MetricData, AttributeTypeValue, State } from "../codeCharta.model"
 import { hierarchy, HierarchyNode } from "d3"
 import { IRootScopeService } from "angular"
 import { CodeMapHelper } from "../util/codeMapHelper"
@@ -63,34 +53,13 @@ export class MetricService implements FilesSelectionSubscriber, BlacklistSubscri
 	}
 
 	public getAttributeTypeByMetric(metricName: string, state: State): AttributeTypeValue {
-		const attributeType = this.getMergedAttributeTypes(state.fileSettings.attributeTypes).find(x => {
-			return _.findKey(x) === metricName
-		})
-
-		if (attributeType) {
-			return attributeType[metricName]
-		}
-		return null
+		return state.fileSettings.attributeTypes.nodes[metricName]
 	}
 
 	private setNewMetricData() {
 		this.metricData = this.calculateMetrics()
 		this.addUnaryMetric()
 		this.notifyMetricDataAdded()
-	}
-
-	private getMergedAttributeTypes(attributeTypes: AttributeTypes): AttributeType[] {
-		const mergedAttributeTypes = [...attributeTypes.nodes]
-
-		mergedAttributeTypes.forEach(nodeAttribute => {
-			attributeTypes.edges.forEach(edgeAttribute => {
-				if (_.findKey(nodeAttribute) !== _.findKey(edgeAttribute)) {
-					mergedAttributeTypes.push(edgeAttribute)
-				}
-			})
-		})
-
-		return mergedAttributeTypes
 	}
 
 	private calculateMetrics(): MetricData[] {